+++ examples/github_gist/plugins/resource_gist/utils.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 )
 
-func tfResourceDataToModel(s string) (*ResourceData, error) {
-	rd := ResourceData{}
+func tfAttributesToModel(s string) (*Attributes, error) {
+	rd := Attributes{}
 	err := json.Unmarshal([]byte(s), &rd)
 	if err != nil {
 		return nil, fmt.Errorf("could not unmarshal JSON resource data: %w", err)